chore: split scannerdefinitions handler into separate files
